Updates for AWS S3 service: - Refactor to include new service in v2
