+++ src/main/resources/crafter/studio/extension/rendering-overlay-context.xml
@@ -38,7 +38,7 @@
         <property name="uiService" ref="uiService"/>
     </bean>
 
-    <bean id="studioS3Controller" class="org.craftercms.studio.controller.rest.v2.aws.S3Controller"/>
+    <bean id="studioS3Controller" class="org.craftercms.studio.controller.rest.v2.aws.AwsS3Controller"/>
 
     <bean id="studio.binaryView" class="org.craftercms.studio.impl.v1.util.spring.mvc.BinaryView">
         <property name="disableCaching" value="${crafter.core.rest.views.json.disableCaching}"/>
